components/ml/src/requirements.txt
absl-py==1.4.0
adagio==0.2.4
alembic==1.9.2
ansi2html==1.8.0
antlr4-python3-runtime==4.9
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
bayesian-optimization==1.4.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
cma==3.2.2
cmdstanpy==1.1.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.33
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.3.0
databricks-cli==0.17.4
debugpy==1.6.6
decorator==5.1.1
Deprecated==1.2.13
distributed==2023.3.0
dm-tree==0.1.8
docker==6.0.1
dtw-python==1.3.0
entrypoints==0.4
ephem==4.1.4
esig==0.9.7
executing==1.2.0
filterpy==1.4.5
Flask==2.1.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.38.0
fs==2.4.16
fsspec==2023.3.0
fugue==0.6.0
fugue-sql-antlr==0.1.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
gluonts==0.12.3
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
HeapDict==1.0.1
hijri-converter==2.2.4
hmmlearn==0.2.8
holidays==0.20
hydra-core==1.3.1
hydra-nevergrad-sweeper @ file:///G:/aldoum/code/vr-gestures/components/ml/deps/hydra-nevergrad-sweeper-1.2.0.tar
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.2
ipykernel==6.21.2
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter-dash==0.4.2
jupyter_client==8.0.3
jupyter_core==5.2.0
keras==2.11.0
keras-self-attention==0.51.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
libclang==15.0.6.1
llvmlite==0.39.1
locket==1.0.0
LunarCalendar==0.0.9
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mlflow==2.1.1
mne==1.3.1
msgpack==1.0.3
nest-asyncio==1.5.6
nevergrad @ git+https://github.com/facebookresearch/nevergrad@521b6c3959fbea05329b0e52d86b6888dd4ed30f
nptyping==1.4.2
numba==0.56.4
numpy==1.22.0
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
orjson==3.8.7
packaging==22.0
pandas==1.5.2
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
plotly==5.13.1
plotly-resampler==0.8.3.1
pmdarima==2.0.2
pooch==1.7.0
prompt-toolkit==3.0.38
prophet==1.1.2
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
Pygments==2.14.0
PyJWT==2.6.0
pykalman==0.9.5
PyMeeus==0.5.12
pyod==1.0.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
qpd==0.2.4
querystring-parser==1.2.4
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
scikit-learn==1.2.0
scikit-posthocs==0.7.0
scipy==1.10.0
seaborn==0.12.2
shap==0.41.0
six==1.16.0
sktime==0.15.1
slicer==0.0.7
smmap==5.0.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.46
sqlglot==11.3.0
sqlparse==0.4.3
stack-data==0.6.2
statsforecast==1.3.0
statsmodels==0.13.5
stumpy==1.11.1
tabulate==0.9.0
tbats==1.1.2
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.19.0
termcolor==2.2.0
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
trace-updater==0.0.9
traitlets==5.9.0
triad==0.8.2
tsfresh==0.20.0
tslearn==0.5.3.2
typing_extensions==4.4.0
typish==1.9.3
urllib3==1.26.14
waitress==2.1.2
wcwidth==0.2.6
websocket-client==1.5.0
Werkzeug==2.1.2
wrapt==1.14.1
xarray==2023.1.0
zict==2.2.0
zipp==3.11.0
